Prerender /dashboard/phone_numbers
The phone numbers section shipped in the dashboard app but was never added to
vite.config.ts's hardcoded prerender list, so the build emitted no HTML for
/dashboard/phone_numbers and Vercel served its own 404 — the router's
redirect-to-workflows fallback never ran, because the app never loaded. A
customer hit this on their first visit to the page.

Derive the prerender list from the section list instead of maintaining both.
dashboardSections moves to a React-free module so vite.config.ts can import
it, which is what kept the two in sync from being possible before.

Verified: dist/client/dashboard/phone_numbers is now emitted by the build.
